# server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "EPS API Documentation",
    "version": "1.0.0",
    "description": "API documentation for the Exam Plagiarism System (EPS). Use this to test the endpoints."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ProjectSubmission": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "submissionUrl": {
            "type": "string",
            "example": "https://github.com/user/repo"
          },
          "status": {
            "type": "string",
            "enum": [
              "submitted",
              "graded",
              "late",
              "needs_revision"
            ],
            "example": "submitted"
          },
          "grade": {
            "type": "integer",
            "nullable": true,
            "example": 95
          },
          "feedback": {
            "type": "string",
            "nullable": true,
            "example": "Great work!"
          },
          "userId": {
            "type": "integer"
          },
          "projectId": {
            "type": "integer"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/projects/update/{id}": {
      "patch": {
        "summary": "Update an existing project by its ID",
        "tags": [
          "Projects"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The ID of the project to update."
          }
        ],
        "requestBody": {
          "description": "Fields to update for the project.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Updated API Project Title"
                  },
                  "description": {
                    "type": "string",
                    "example": "Updated project description."
                  },
                  "deadline": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2025-01-10T23:59:00Z"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Project updated successfully."
          },
          "400": {
            "description": "Bad request."
          },
          "401": {
            "description": "Unauthorized."
          },
          "404": {
            "description": "Project not found."
          }
        }
      }
    },
    "/api/projects/delete/{id}": {
      "delete": {
        "summary": "Delete a project by its ID",
        "tags": [
          "Projects"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The ID of the project to delete."
          }
        ],
        "responses": {
          "200": {
            "description": "Project deleted successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "404": {
            "description": "Project not found."
          }
        }
      }
    },
    "/api/project-submissions/create": {
      "post": {
        "summary": "Create a new project submission",
        "tags": [
          "Project Submissions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "projectId",
                  "submissionUrl"
                ],
                "properties": {
                  "projectId": {
                    "type": "integer",
                    "description": "The ID of the project being submitted.",
                    "example": 1
                  },
                  "submissionUrl": {
                    "type": "string",
                    "description": "The URL to the student's work (e.g., GitHub, Google Drive).",
                    "example": "https://github.com/student/my-project-repo"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Submission created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectSubmission"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (e.g., missing fields or submission already exists)."
          },
          "401": {
            "description": "Unauthorized."
          },
          "404": {
            "description": "Project not found."
          }
        }
      }
    },
    "/api/project-submissions/all": {
      "get": {
        "summary": "Get all submissions for the authenticated user",
        "tags": [
          "Project Submissions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "An array of the user's project submissions.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProjectSubmission"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          }
        }
      }
    },
...
    "/api/project-submissions/update/{id}": {
      "patch": {
        "summary": "Update a specific project submission",
        "tags": [
          "Project Submissions"
        ],
        "description": "Allows a user to update their own submission (e.g., change the URL).",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The ID of the submission to update."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "submissionUrl": {
                    "type": "string",
                    "description": "The new URL for the submission.",
                    "example": "https://github.com/student/my-updated-repo"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Submission updated successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden (user trying to update another user's submission)."
          },
          "404": {
            "description": "Submission not found."
          }
        }
      }
    },
    "/api/project-submissions/delete/{id}": {
      "delete": {
        "summary": "Delete a project submission",
        "tags": [
          "Project Submissions"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The ID of the submission to delete."
          }
        ],
        "responses": {
          "200": {
            "description": "Submission deleted successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Submission not found."
          }
        }
      }
    },
    "/api/project-submissions/submissions/{id}/status": {
      "put": {
        "summary": "Update the status of a submission (Instructor/Admin)",
        "tags": [
          "Project Submissions"
        ],
        "description": "Used by instructors or admins to grade or change the status of a submission.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The ID of the submission to update."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "submitted",
                      "graded",
                      "needs_revision",
                      "accepted"
                    ],
                    "example": "graded"
                  },
                  "grade": {
                    "type": "integer",
                    "example": 92
                  },
                  "feedback": {
                    "type": "string",
                    "example": "Well done, but please add more comments to your code."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Submission status updated successfully."
          },
          "400": {
            "description": "Bad request (e.g., invalid status value)."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden (e.g., a student trying to grade a submission)."
          },
          "404": {
            "description": "Submission not found."
          }
        }
      }
    },
...
  },
...
}